Add MIT License to the project
An MIT License file was added to the project. This license gives users express permission to use, modify, distribute and sell the software freely. It also disclaims any warranties and states the conditions for reusing the software. This step is crucial in securing mutual understanding and agreement between the developers and users of the software.